requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and Storage
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI and Machine Learning
torch==2.1.1
transformers==4.36.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.4

# ESM3 and Biological Modeling
fair-esm==2.0.0
biotite==0.39.0
biopython==1.81

# HTTP Client and Async
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.0

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Distributed Computing (ExoStack dependencies)
celery==5.3.4
kombu==5.3.4
docker==6.1.3

# Additional utilities
python-dateutil==2.8.2
cryptography==41.0.8
jinja2==3.1.2
